Fix build for 64 bit Windows
This just fixes some casts regarding handles. We use intprtr_t
as a generic type for HANDLES to avoid ifdefs even if HANDLE
values can be represented in 32bit also on Win64.

+++ src/uiserver/assuanserverconnection.cpp
@@ -220,7 +220,7 @@ Q_SIGNALS:
     void startKeyManager();
 
 public Q_SLOTS:
-    void slotReadActivity(int)
+    void slotReadActivity(intptr_t)
     {
         Q_ASSERT(ctx);
         int done = false;
